agent_ui: Move UI logic from `ThreadHistory` to separate component
